feat: Enable an item to be added to a list
An alert view will be presented when the 'add' button is tapped
and it will have a text field.  Text entered here will be included
in the data model, or the user can cancel and leave the list
unmodified.  The table displays the data that is currently in the
data model.

No data persistence (yet).